+++ packages/react-native-sdk/src/components/Call/CallLayout/CallParticipantsGrid.tsx
@@ -2,21 +2,9 @@ import React from 'react';
 import { StyleSheet, View } from 'react-native';
 import { useCallStateHooks } from '@stream-io/video-react-bindings';
 import { useDebouncedValue } from '../../../utils/hooks/useDebouncedValue';
-import {
-  CallParticipantsList,
-  CallParticipantsListProps,
-} from '../CallParticipantsList/CallParticipantsList';
+import { CallParticipantsListProps } from '../CallParticipantsList/CallParticipantsList';
 import { ComponentTestIds } from '../../../constants/TestIds';
-import {
-  ParticipantNetworkQualityIndicator as DefaultParticipantNetworkQualityIndicator,
-  ParticipantReaction as DefaultParticipantReaction,
-  ParticipantLabel as DefaultParticipantLabel,
-  ParticipantVideoFallback as DefaultParticipantVideoFallback,
-  VideoRenderer as DefaultVideoRenderer,
-  ParticipantView as DefaultParticipantView,
-  LocalParticipantView as DefaultLocalParticipantView,
-  LocalParticipantViewProps,
-} from '../../Participant';
+import { LocalParticipantViewProps } from '../../Participant';
 
 /**
  * Props for the CallParticipantsGrid component.
@@ -31,51 +19,57 @@ export type CallParticipantsGridProps = Pick<
   | 'VideoRenderer'
 > & {
   /**
-   * Component to customize the local participant view.
+   * Component to customize the LocalParticipantView.
    */
   LocalParticipantView?: React.ComponentType<LocalParticipantViewProps>;
+  /**
+   * Component to customize the CallParticipantsList.
+   */
+  CallParticipantsList?: React.ComponentType<CallParticipantsListProps>;
 };
 
 /**
  * Component used to display the list of participants in a grid mode.
  */
 export const CallParticipantsGrid = ({
-  ParticipantLabel = DefaultParticipantLabel,
-  ParticipantNetworkQualityIndicator = DefaultParticipantNetworkQualityIndicator,
-  ParticipantReaction = DefaultParticipantReaction,
-  ParticipantVideoFallback = DefaultParticipantVideoFallback,
-  ParticipantView = DefaultParticipantView,
-  VideoRenderer = DefaultVideoRenderer,
-  LocalParticipantView = DefaultLocalParticipantView,
+  CallParticipantsList,
+  ParticipantLabel,
+  ParticipantNetworkQualityIndicator,
+  ParticipantReaction,
+  ParticipantVideoFallback,
+  ParticipantView,
+  VideoRenderer,
+  LocalParticipantView,
 }: CallParticipantsGridProps) => {
   const { useRemoteParticipants, useParticipants } = useCallStateHooks();
   const _remoteParticipants = useRemoteParticipants();
   const allParticipants = useParticipants();
   const remoteParticipants = useDebouncedValue(_remoteParticipants, 300); // we debounce the remote participants to avoid unnecessary rerenders that happen when participant tracks are all subscribed simultaneously
 
-  const showFloatingView = remoteParticipants.length < 3;
-  const isUserAloneInCall = remoteParticipants?.length === 0;
-  const participants = showFloatingView ? remoteParticipants : allParticipants;
+  const showFloatingView =
+    remoteParticipants.length > 0 && remoteParticipants.length < 3;
 
-  if (showFloatingView && isUserAloneInCall) {
-    return <LocalParticipantView layout={'fullscreen'} />;
-  }
+  const participants = showFloatingView ? remoteParticipants : allParticipants;
 
   return (
     <View
       style={styles.container}
       testID={ComponentTestIds.CALL_PARTICIPANTS_GRID}
     >
-      {showFloatingView && <LocalParticipantView layout={'floating'} />}
-      <CallParticipantsList
-        participants={participants}
-        ParticipantLabel={ParticipantLabel}
-        ParticipantNetworkQualityIndicator={ParticipantNetworkQualityIndicator}
-        ParticipantReaction={ParticipantReaction}
-        ParticipantVideoFallback={ParticipantVideoFallback}
-        ParticipantView={ParticipantView}
-        VideoRenderer={VideoRenderer}
-      />
+      {showFloatingView && LocalParticipantView && <LocalParticipantView />}
+      {CallParticipantsList && (
+        <CallParticipantsList
+          participants={participants}
+          ParticipantLabel={ParticipantLabel}
+          ParticipantNetworkQualityIndicator={
+            ParticipantNetworkQualityIndicator
+          }
+          ParticipantReaction={ParticipantReaction}
+          ParticipantVideoFallback={ParticipantVideoFallback}
+          ParticipantView={ParticipantView}
+          VideoRenderer={VideoRenderer}
+        />
+      )}
     </View>
   );
 };